make tskserver port selection dynamic
If multiple tskservers are started concurrently on the same host,
only one of them will be able to connect to the (hardcoded) tskserver
port.  This patch allows dynamic port selection by searching for
a monotonically increasing port number for each tskserver instance.
This is highly desirable for installations that do not have whole
node allocation policies.